requirements.txt
openai>=1.14.0
python-dotenv>=1.0.0
ddgs>=7.0.0
mcp>=1.0.0
mcp-atlassian>=0.21.0
